refactor(platform): split platform/api/views.py into domain modules
Closes #641
