+++ django_email_learning/platform/api/views/__init__.py
@@ -1,0 +1,93 @@
+from django_email_learning.platform.api.views.courses import (
+    CourseCreationMixin,
+    CourseView,
+    CourseContentView,
+    ReorderCourseContentView,
+    SingleCourseContentView,
+    SingleCourseView,
+    SendLessonToPlatformUser,
+)
+from django_email_learning.platform.api.views.oauth import (
+    OauthSessionView,
+    OauthGetGroupListView,
+    OauthGroupEnrollment,
+    ImapConnectionView,
+)
+from django_email_learning.platform.api.views.organisations import (
+    OrganizationsView,
+    OrganizationUsersView,
+    SingleOrganizationUserView,
+    SingleOrganizationView,
+    GetOrCreateUserByEmail,
+)
+from django_email_learning.platform.api.views.assignments import (
+    SubmittedAssignmentsView,
+    SubmittedAssignmentDetailView,
+    SubmissionReview,
+)
+from django_email_learning.platform.api.views.learners import (
+    LearnersView,
+    SingleLearnerView,
+    EnrollmentsView,
+    EnrollmentView,
+    EnrollmentsStatisticsView,
+)
+from django_email_learning.platform.api.views.newsletters import (
+    NewsletterView,
+    SingleNewsletterView,
+    SendoutView,
+    SingleSendoutView,
+    SubscriberView,
+    SingleSubscriberView,
+    SubscribersCsvExportView,
+)
+from django_email_learning.platform.api.views.misc import (
+    JobHealthStatus,
+    ApiKeyView,
+    SingleApiKeyView,
+    JobsStatus,
+    RootView,
+    FileView,
+    UpdateSessionView,
+)
+
+__all__ = [
+    "CourseCreationMixin",
+    "CourseView",
+    "CourseContentView",
+    "ReorderCourseContentView",
+    "SingleCourseContentView",
+    "SingleCourseView",
+    "SendLessonToPlatformUser",
+    "OauthSessionView",
+    "OauthGetGroupListView",
+    "OauthGroupEnrollment",
+    "ImapConnectionView",
+    "OrganizationsView",
+    "OrganizationUsersView",
+    "SingleOrganizationUserView",
+    "SingleOrganizationView",
+    "GetOrCreateUserByEmail",
+    "SubmittedAssignmentsView",
+    "SubmittedAssignmentDetailView",
+    "SubmissionReview",
+    "LearnersView",
+    "SingleLearnerView",
+    "EnrollmentsView",
+    "EnrollmentView",
+    "EnrollmentsStatisticsView",
+    "NewsletterView",
+    "SingleNewsletterView",
+    "SendoutView",
+    "SingleSendoutView",
+    "SubscriberView",
+    "SingleSubscriberView",
+    "SubscribersCsvExportView",
+    "JobHealthStatus",
+    "ApiKeyView",
+    "SingleApiKeyView",
+    "JobsStatus",
+    "RootView",
+    "FileView",
+    "UpdateSessionView",
+]
